searchengine/searchengine/requirements.txt
# Core Django Framework
Django==4.2.22
gunicorn==23.0.0

# AI/ML Libraries
sentence-transformers==3.4.1
torch==2.6.0
faiss-cpu==1.10.0

# Data Processing
numpy>=1.26.4,<2.0.0
pandas==2.2.3

# HTTP & Utilities
requests==2.32.3
psutil==5.9.8

# Environment Management
python-dotenv==1.0.1

# ML Dependencies (required by sentence-transformers)
transformers==4.50.0
scikit-learn==1.6.1
scipy==1.13.1
huggingface-hub==0.29.3
tokenizers==0.21.1
safetensors==0.5.3
tqdm==4.67.1

# Django Dependencies
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2

# Common Dependencies
certifi==2025.1.31
urllib3==2.3.0
idna==3.10
charset-normalizer==3.4.1
packaging==24.2
filelock==3.18.0
fsspec==2025.3.0
PyYAML==6.0.2
regex==2024.11.6
python-dateutil==2.9.0.post0
pytz==2025.1
joblib==1.4.2
threadpoolctl==3.6.0
typing_extensions==4.12.2
setuptools==78.1.0
wheel==0.45.1
six==1.17.0
